// src/constants/enum.ts
export enum Role {
  Employee = 'Employee',
  Owner = 'Owner',
  Guest = 'Guest',
}

export enum DishStatus {
  Available = 'Available',
  Unavailable = 'Unavailable',
  Hidden = 'Hidden',
}

export enum OrderStatus {
  Pending = 'Pending',
  Processing = 'Processing',
  Rejected = 'Rejected',
  Delivered = 'Delivered',
  Paid = 'Paid',
}

export enum TableStatus {
  Available = 'Available',
  Hidden = 'Hidden',
  Reserved = 'Reserved',
}